// src/permissions/enums/permission-permissions.enum.ts
export enum permissionPermissionsEnum {
  CREATE = "ساخت دسترسی ",
  LIST = "دریافت لیست دسترسی ها ",
  FIND_ONE = "دریافت یک لیست دسترسی ",
  UPDATE = "به روز رسانی دسترسی ها",
  DELETE = "حذف دسترسی",
  ADD_ROLE = "اضافه کردن دسترسی به نقش",
  REMOVE_ROLE = "حذف دسترسی از نقش",
  ACCESS_CHECK = "بررسی دسترسی",
  HARD_DELETE = "حذف عمیق دسترسی",
}